pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "vllm-chunked-vision-encoder"
version = "0.1.0"
description = "Chunked vision ingestion, dynamic multimodal budgeting, and MCP integration for vLLM."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "OpenAI Codex" }]
keywords = ["vllm", "vision", "vlm", "mcp", "inference"]
dependencies = [
  "httpx>=0.28.0",
  "numpy>=1.26.0",
  "Pillow>=10.4.0",
  "psutil>=5.9.8",
  "torch>=2.5.0",
  "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.13.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]
hf = [
  "torchvision>=0.20.0",
  "transformers>=4.48.0",
]
server = [
  "mcp>=1.26.0",
  "uvicorn>=0.34.0",
]
vllm = [
  "vllm>=0.8.0",
]
all = [
  "mcp>=1.26.0",
  "torchvision>=0.20.0",
  "transformers>=4.48.0",
  "uvicorn>=0.34.0",
  "vllm>=0.8.0",
]

[project.scripts]
vllm-chunked-vision-mcp = "vllm_chunked_vision.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["vllm_chunked_vision"]

[tool.mypy]
python_version = "3.11"
packages = ["vllm_chunked_vision"]
strict = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
